Fixed issus: 1. ICMP some bugs 2. cgnat_interface_add_del not worker barrier 3. TCP state rollback 4. pre_alloc mode some bugs 5. cgnat_delete_user inaccurate active users 6. tcp mss out2in

+++ ET2500/vpp-24.02/src/plugins/nat/cgnat/cgnat.h
@@ -273,7 +273,8 @@ typedef struct
   /* ALLOCATED only. COOLING is tracked separately and is not reusable. */
   u16 allocated_blocks;
   u16 cooling_blocks;
-  u16 active_users;
+  /* Updated atomically: user deletion paths do not always hold ip->lock. */
+  u32 active_users;
 
   clib_spinlock_t lock;
